projetIAAgent/requirements.txt
fastapi
uvicorn
Flask
Flask-Login
Flask-WTF
bcrypt
SQLAlchemy
pypdf2
pdfplumber
pytesseract
python-docx
anthropic
rich
openpyxl
pytest
python-multipart
jinja2
python-dotenv
aiohttp
